requirements.txt
matplotlib==3.10.3
numpy==1.24.3
open3d==0.19.0
opencv_python==4.10.0.84
PyYAML==6.0.1
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.15.3
tqdm==4.67.1
tqdm==4.66.4
tqdm==4.66.5
tqdm==4.65.0
